+++ module-order/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+    implementation project(':module-menu')
+    implementation project(':module-table')
+    implementation project(':module-tablegroup')
+    testFixturesApi(testFixtures(project(':module-table')))
+    testFixturesApi(testFixtures(project(':module-menu')))
+}
