+++ module-order/src/main/java/kitchenpos/order/domain/TableEmptyChangedEventHandler.java
