+++ module-order/src/test/java/kitchenpos/OrderApplicationTests.java
@@ -1,0 +1,12 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OrderApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+}
