+++ module-table/build.gradle
@@ -1,6 +1,4 @@
 dependencies {
-//    implementation project(':module-product')
-//    implementation project(':module-menugroup')
     testFixturesApi(testFixtures(project(':module-menu')))
     testFixturesApi(testFixtures(project(':module-order')))
 }
